CUDA vector types float2/3/4 and int2/3/4 are now exported to Lua.
Lua scene file format figured out. It was a long roundabout way to get
there, ditching LuaBind, ditching userdata, and eventually settling
on raw tables, but it's concise and works nicely.

It's not completely fleshed out, but it has a sphere and is ready for
more fleshing out as we add object types.

The scene class has been refactored to some globals inside the scene
namespace to facilitate easy calling from the Lua callbacks.

The scene::Insert() method is a function template that should work for
whatever new object types we throw at it. It takes in the enum ObjType
and a pointer to that object. It allocates space at the end of the object
chunk and dumps its data in there and allocates space at the end of the
meta chunk for a new meta object and sets up its id and type. It also
sets the metaobject's pointer to the location in the host-allocated
object chunk where that object lives. (In other words, it can be
traversed safely on the host). Before pushing the scene data up to the
device, we need to traverse the metaobjects and translate the host
pointers into device pointers (by calculating the offset from the host
chunk base pointer, then adding this offset to the device chunk base
pointer).

Commit messages from 15 squashed commits, in chronological order:

    * Big refactoring, moving things around. Names are more sensical now.
      Prepping for more LuaBind work.

    * A little more refactoring...

    * Starting work on binding surface_t.

    * Updating premake script to compile CUDA code with minimum compute
      capability of 1.1.

    * Work on export surface. LuaBind crashing on registration. Gonna try
      something to fix it.

    * Surface working now. Trying to figure out Sphere.

    * Experiment worked, LuaBind removed. Writing interface code myself.

    * Got int2/3/4 working along with float3. Just have to do float2/4 and
      vectors will be finished.

    * Fixed some boo-boos with the int2/3/4 types inside the new functions.

    * Finished the Lua bindings to CUDA vector types, although I think I'm
      going to end up throwing these out. Yay.

    * New Lua extraction method working. Vectors, surface, and sphere done
      so far.

    * Basic macros up and rolling. Extracted defaults out to defaults.h.
      Render macro now sets global.

    * Working on Lua macros to dump object into flat memory chunks.

    * Function template for inserting objects.